+++ app/src/main/java/com/netchar/nicknamer/presentation/infrastructure/Utils.kt
@@ -30,9 +30,9 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.DefaultLifecycleObserver
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleOwner
+import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
 import com.netchar.nicknamer.R
-import com.netchar.nicknamer.presentation.infrastructure.helpers.BindableViewHolder
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 
@@ -63,8 +63,8 @@ fun View.visible(visible: Boolean) {
     this.visibility = if (visible) View.VISIBLE else View.GONE
 }
 
-fun <T : BindableViewHolder<*, *>> T.listen(event: (position: Int) -> Unit): T {
-    binding.root.setOnClickListener {
+fun <T : RecyclerView.ViewHolder> T.listen(event: (position: Int) -> Unit): T {
+    itemView.setOnClickListener {
         event.invoke(adapterPosition)
     }
     return this
